Move Multi Dimension indexation from Framework to InventoryMultiDimensionIndex modules
