+++ app/code/Magento/InventoryCache/Model/ResourceModel/GetProductIdsByStockIds.php
@@ -10,10 +10,10 @@ namespace Magento\InventoryCache\Model\ResourceModel;
 use Magento\Catalog\Api\Data\ProductInterface;
 use Magento\Framework\App\ResourceConnection;
 use Magento\Framework\EntityManager\MetadataPool;
-use Magento\Framework\MultiDimensionalIndexer\Alias;
-use Magento\Framework\MultiDimensionalIndexer\IndexName;
-use Magento\Framework\MultiDimensionalIndexer\IndexNameBuilder;
-use Magento\Framework\MultiDimensionalIndexer\IndexNameResolverInterface;
+use Magento\InventoryMultiDimensionalIndexer\Model\Alias;
+use Magento\InventoryMultiDimensionalIndexer\Model\IndexName;
+use Magento\InventoryMultiDimensionalIndexer\Model\IndexNameBuilder;
+use Magento\InventoryMultiDimensionalIndexer\Model\IndexNameResolverInterface;
 use Magento\InventoryCatalogApi\Api\DefaultStockProviderInterface;
 use Magento\InventoryIndexer\Indexer\IndexStructure;
 use Magento\InventoryIndexer\Indexer\InventoryIndexer;
