+++ app/code/Magento/InventoryCache/composer.json
@@ -4,6 +4,7 @@
     "require": {
         "php": "~7.1.3||~7.2.0",
         "magento/framework": "*",
+        "magento/module-inventory-multi-dimensional-indexer": "*",
         "magento/module-catalog": "*",
         "magento/module-inventory": "*",
         "magento/module-inventory-catalog-api": "*",
