+++ app/code/Magento/InventoryConfigurableProductIndexer/Indexer/Stock/StockIndexer.php
@@ -9,11 +9,11 @@ namespace Magento\InventoryConfigurableProductIndexer\Indexer\Stock;
 
 use Magento\Framework\App\ResourceConnection;
 use Magento\Framework\Exception\StateException;
-use Magento\Framework\MultiDimensionalIndexer\Alias;
-use Magento\Framework\MultiDimensionalIndexer\IndexHandlerInterface;
-use Magento\Framework\MultiDimensionalIndexer\IndexNameBuilder;
-use Magento\Framework\MultiDimensionalIndexer\IndexStructureInterface;
-use Magento\Framework\MultiDimensionalIndexer\IndexTableSwitcherInterface;
+use Magento\InventoryMultiDimensionalIndexer\Model\Alias;
+use Magento\InventoryMultiDimensionalIndexer\Model\IndexHandlerInterface;
+use Magento\InventoryMultiDimensionalIndexer\Model\IndexNameBuilder;
+use Magento\InventoryMultiDimensionalIndexer\Model\IndexStructureInterface;
+use Magento\InventoryMultiDimensionalIndexer\Model\IndexTableSwitcherInterface;
 use Magento\InventoryCatalogApi\Api\DefaultStockProviderInterface;
 use Magento\InventoryIndexer\Indexer\InventoryIndexer;
 use Magento\InventoryIndexer\Indexer\Stock\GetAllStockIds;
