+++ app/code/Magento/InventoryIndexer/Indexer/SourceItem/SourceItemIndexer.php
@@ -8,10 +8,10 @@ declare(strict_types=1);
 namespace Magento\InventoryIndexer\Indexer\SourceItem;
 
 use Magento\Framework\App\ResourceConnection;
-use Magento\Framework\MultiDimensionalIndexer\Alias;
-use Magento\Framework\MultiDimensionalIndexer\IndexHandlerInterface;
-use Magento\Framework\MultiDimensionalIndexer\IndexNameBuilder;
-use Magento\Framework\MultiDimensionalIndexer\IndexStructureInterface;
+use Magento\InventoryMultiDimensionalIndexer\Model\Alias;
+use Magento\InventoryMultiDimensionalIndexer\Model\IndexHandlerInterface;
+use Magento\InventoryMultiDimensionalIndexer\Model\IndexNameBuilder;
+use Magento\InventoryMultiDimensionalIndexer\Model\IndexStructureInterface;
 use Magento\InventoryCatalogApi\Api\DefaultStockProviderInterface;
 use Magento\InventoryIndexer\Indexer\InventoryIndexer;
 use Magento\InventoryIndexer\Indexer\Stock\StockIndexer;
