+++ app/code/Magento/InventoryIndexer/Model/StockIndexTableNameResolver.php
@@ -8,9 +8,9 @@ declare(strict_types=1);
 namespace Magento\InventoryIndexer\Model;
 
 use Magento\Framework\App\ResourceConnection;
-use Magento\Framework\MultiDimensionalIndexer\Alias;
-use Magento\Framework\MultiDimensionalIndexer\IndexNameBuilder;
-use Magento\Framework\MultiDimensionalIndexer\IndexNameResolver;
+use Magento\InventoryMultiDimensionalIndexer\Model\Alias;
+use Magento\InventoryMultiDimensionalIndexer\Model\IndexNameBuilder;
+use Magento\InventoryMultiDimensionalIndexer\Model\IndexNameResolver;
 use Magento\InventoryIndexer\Indexer\InventoryIndexer;
 
 /**
