+++ app/code/Magento/InventoryMultiDimensionalIndexer/Model/IndexNameBuilder.php
@@ -5,7 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Magento\Framework\MultiDimensionalIndexer;
+namespace Magento\InventoryMultiDimensionalIndexer\Model;
 
 use Magento\Framework\ObjectManagerInterface;
