+++ app/code/Magento/InventoryMultiDimensionalIndexer/README.md
@@ -1,6 +1,10 @@
+# Inventory
+
+**InventoryMultiDimensionalIndexer**
+
 MultiDimensionalIndexer
 =======
-The **\Magento\Framework\MultiDimensionalIndexer** library provides functionality of multi-dimension index creation and
+The InventoryMultiDimensionalIndexer provides functionality of multi-dimension index creation and
 handling.
 
 Library introduces a set of extension points which split monolithic index by specified Dimension (Scope), creating 
@@ -9,3 +13,4 @@ resolving mechanism based on provided scope. The Multi-Dimension indexes introdu
 and ability to reindex data in the scope of particular Dimension only.
 
 Aliasing mechanism guarantees zero downtime to make Front-End responsive while Full Reindex being processed.
+
