+++ app/code/Magento/InventoryMultiDimensionalIndexer/composer.json
@@ -1,0 +1,21 @@
+{
+  "name": "magento/module-inventory-multi-dimensional-indexer",
+  "description": "N/A",
+  "require": {
+    "php": "~7.1.3||~7.2.0",
+    "magento/framework": "*"
+  },
+  "type": "magento2-module",
+  "license": [
+    "OSL-3.0",
+    "AFL-3.0"
+  ],
+  "autoload": {
+    "files": [
+      "registration.php"
+    ],
+    "psr-4": {
+      "Magento\\InventoryMultiDimensionalIndexer\\": ""
+    }
+  }
+}
