+++ app/code/Magento/InventoryMultiDimensionalIndexer/etc/di.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
+    <type name="Magento\InventoryMultiDimensionalIndexer\Model\IndexNameResolver">
+        <arguments>
+            <argument name="indexScopeResolver" xsi:type="object">Magento\Framework\Indexer\ScopeResolver\IndexScopeResolver</argument>
+        </arguments>
+    </type>
+    <preference for="Magento\InventoryMultiDimensionalIndexer\Model\IndexTableSwitcherInterface" type="Magento\InventoryMultiDimensionalIndexer\Model\IndexTableSwitcher" />
+    <preference for="Magento\InventoryMultiDimensionalIndexer\Model\IndexNameResolverInterface" type="Magento\InventoryMultiDimensionalIndexer\Model\IndexNameResolver" />
+
+</config>
+
+
